refactor: split test cases and package dependencies for core
- Replace requires in "artillery" to load the new internal "core" package
- Update require paths inside "core"
- Update require paths in tests
- Move some test cases from "core" back into "artillery" package
