+++ .dockerignore
@@ -3,7 +3,6 @@
 
 # Explicitly include only the necessary files we want in the Docker image.
 !bin
-!core
 !lib
 !LICENSE.txt
 !README.md
