+++ packages/artillery/package.json
@@ -46,7 +46,7 @@
     }
   },
   "scripts": {
-    "test": "export ARTILLERY_TELEMETRY_DEFAULTS='{\"source\":\"test-suite\"}' && tap --no-coverage --timeout=300 test/cli/probe/*.js test/cli/*.test.js && bash test/runner.sh && bash test/lib/run.sh && node test/testcases/plugins/*.test.js",
+    "test": "export ARTILLERY_TELEMETRY_DEFAULTS='{\"source\":\"test-suite\"}' && tap --no-coverage --timeout=300 test/unit/*.js test/cli/probe/*.js test/cli/*.test.js && bash test/runner.sh && bash test/lib/run.sh && node test/testcases/plugins/*.test.js",
     "is_linted": "find . -name '*.js' | grep -v node_modules | xargs eslint",
     "lint": "eslint --ext \".js,.ts,.tsx\" --ignore-path .gitignore .",
     "lint-fix": "npm run lint -- --fix",
@@ -84,86 +84,60 @@
     "email": "npm@veldstra.org"
   },
   "dependencies": {
-    "@artilleryio/sketches-js": "^1.0.4",
     "@aws-sdk/credential-providers": "^3.127.0",
     "@oclif/command": "^1.8.16",
     "@oclif/config": "^1.18.3",
+    "@oclif/dev-cli": "^1.22.2",
     "@oclif/plugin-help": "^3.3.1",
-    "agentkeepalive": "^4.1.0",
     "archiver": "^5.3.1",
-    "arrivals": "^2.1.2",
     "artillery-plugin-ensure": "^1.1.1",
     "artillery-plugin-expect": "^2.1.0",
     "artillery-plugin-metrics-by-endpoint": "^1.0.2",
     "artillery-plugin-publish-metrics": "^2.0.2",
     "async": "^2.6.4",
     "aws-sdk": "^2.1172.0",
+    "bats": "^1.6.0",
     "chalk": "^2.4.2",
     "cheerio": "^1.0.0-rc.10",
     "ci-info": "^2.0.0",
     "cli-highlight": "^2.1.11",
     "cli-table3": "^0.6.2",
-    "csv-parse": "^4.16.3",
     "core": "*",
+    "csv-parse": "^4.16.3",
+    "commons": "*",
     "debug": "^4.3.1",
-    "decompress-response": "^6.0.0",
-    "deep-for-each": "^3.0.0",
     "dotenv": "^16.0.1",
-    "driftless": "^2.0.3",
-    "esprima": "^4.0.0",
     "eventemitter3": "^4.0.4",
-    "fast-deep-equal": "^3.1.3",
-    "filtrex": "^0.5.4",
-    "form-data": "^3.0.0",
     "fs-extra": "^10.1.0",
-    "got": "^11.8.5",
-    "hpagent": "^0.1.1",
-    "https-proxy-agent": "^5.0.0",
     "ip": "^1.1.8",
     "jmespath": "^0.16.0",
     "joi": "^17.6.0",
     "js-yaml": "^3.13.1",
-    "jsonpath": "^1.1.1",
     "lodash": "^4.17.19",
     "mime-types": "^2.1.35",
     "moment": "^2.29.4",
-    "nanotimer": "^0.3.15",
-    "oclif": "^3.1.0",
     "ora": "^4.0.4",
     "posthog-node": "^1.3.0",
-    "socket.io-client": "^4.5.1",
-    "socketio-wildcard": "^2.0.0",
     "sprintf-js": "^1.1.2",
     "sqs-consumer": "^5.7.0",
     "temp": "^0.9.4",
     "tmp": "0.2.1",
-    "tough-cookie": "^4.0.0",
     "try-require": "1.2.1",
-    "update-notifier": "^5.1.0",
-    "uuid": "^8.0.0",
-    "ws": "^7.5.7"
+    "update-notifier": "^5.1.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^9.1.2",
     "@commitlint/config-conventional": "^7.1.2",
-    "@hapi/basic": "^6.0.0",
     "@hapi/hapi": "^20.1.3",
-    "@oclif/dev-cli": "^1.22.2",
-    "bats": "^1.6.0",
-    "cookie-parser": "^1.4.3",
     "eslint": "^8.6.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "execa": "^0.10.0",
-    "express": "^4.16.3",
     "get-bin-path": "^5.1.0",
     "husky": "^4.2.5",
-    "nock": "^11.8.2",
     "prettier": "^2.5.1",
-    "proxy": "^1.0.2",
     "rewiremock": "^3.14.3",
     "sinon": "^4.5.0",
-    "socket.io": "^3.1.2",
     "tap": "^15.2.3",
     "zx": "^4.2.0"
   },
