+++ packages/artillery/test/runner.sh
@@ -1,7 +1,5 @@
 #!/usr/bin/env bash
 
-# set -eu -o pipefail
-## && bash test/core/run.sh
 echo "# starting target"
 # &>/dev/null PORT=3003 node ./test/core/targets/simple.js &
 target_pid=$!
