+++ packages/artillery/test/scripts/http/simple_function.json
