+++ packages/commons/jitter.js
