+++ packages/core/package.json
@@ -1,4 +1,49 @@
 {
   "name": "core",
-  "main": "index.js"
+  "main": "index.js",
+  "dependencies": {
+    "@artilleryio/sketches-js": "^1.0.4",
+    "@hapi/basic": "^6.0.0",
+    "agentkeepalive": "^4.1.0",
+    "arrivals": "^2.1.2",
+    "async": "^2.6.4",
+    "chalk": "^2.4.2",
+    "cheerio": "^1.0.0-rc.10",
+    "cookie-parser": "^1.4.3",
+    "csv-parse": "^4.16.3",
+    "commons": "*",
+    "debug": "^4.3.1",
+    "decompress-response": "^6.0.0",
+    "deep-for-each": "^3.0.0",
+    "driftless": "^2.0.3",
+    "esprima": "^4.0.0",
+    "eventemitter3": "^4.0.4",
+    "express": "^4.16.3",
+    "fast-deep-equal": "^3.1.3",
+    "filtrex": "^0.5.4",
+    "form-data": "^3.0.0",
+    "got": "^11.8.5",
+    "hpagent": "^0.1.1",
+    "https-proxy-agent": "^5.0.0",
+    "jsonpath": "^1.1.1",
+    "lodash": "^4.17.19",
+    "nock": "^11.8.2",
+    "proxy": "^1.0.2",
+    "socket.io": "^3.1.2",
+    "socket.io-client": "^4.5.1",
+    "socketio-wildcard": "^2.0.0",
+    "tap": "^15.2.3",
+    "tough-cookie": "^4.0.0",
+    "try-require": "^1.2.1",
+    "uuid": "^8.0.0",
+    "ws": "^7.5.7"
+  },
+  "scripts": {
+    "test": "tap --no-coverage --timeout=300 test/core/unit/*.test.js && bash test/core/run.sh"
+  },
+  "devDependencies": {
+    "@hapi/hapi": "^20.1.3",
+    "rewiremock": "^3.14.3",
+    "sinon": "^4.5.0"
+  }
 }
