+++ packages/core/test/core/test_basic_auth.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const { test } = require('tap');
-const runner = require('../../core').runner;
-const { SSMS } = require('../../core/lib/ssms');
+const runner = require('../..').runner.runner;
+const { SSMS } = require('../../lib/ssms');
 
 test('HTTP basic auth', (t) => {
   const script = require('./scripts/hello_basic_auth.json');
