+++ packages/core/test/core/test_environments.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var { test } = require('tap');
-var runner = require('../../core/lib/runner').runner;
+var runner = require('../../lib/runner').runner;
 var createTarget = require('./lib/interfakify').create;
 var url = require('url');
