+++ packages/core/test/core/test_if.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const { test } = require('tap');
-const runner = require('../../core').runner;
+const runner = require('../..').runner.runner;
 const L = require('lodash');
-const { SSMS } = require('../../core/lib/ssms');
+const { SSMS } = require('../../lib/ssms');
 
 test('ifTrue', (t) => {
   const script = require('./scripts/iftrue.json');
